Add migrations / abstract country list to api to prevent import into clientside
