+++ packages/prisma/schema.prisma
@@ -198,6 +198,7 @@ model User {
   routingForms         App_RoutingForms_Form[] @relation("routing-form")
   verifiedNumbers      VerifiedNumber[]
   hosts                Host[]
+  country              String                  @default("")
 
   @@map(name: "users")
 }
